refactor(detect_card): :truck: rename source files to more appropriate name
