+++ app/detect_card/tests/test_detect_card.py
@@ -1,4 +1,4 @@
-from detect_card import (
+from overlord_card_match import (
     extract_card_text,
     OverlordCardsKeywordsMatcher,
     encode_image,
